:tada: change app.rb classes to modules
